+++ branches/SIMON/SIMON_MINA/simon/src/de/root1/simon/codec/messages/MsgOpenRawChannel.java
@@ -29,18 +29,18 @@ public class MsgOpenRawChannel extends AbstractMessage {
 	
     private static final long serialVersionUID = 1L;
 
-    private byte channelToken;
+    private int channelToken;
     
     public MsgOpenRawChannel() {
     	super(SimonMessageConstants.MSG_OPEN_RAW_CHANNEL);
     }
 
-    public byte getChannelToken() {
+    public int getChannelToken() {
         return channelToken;
     }
 
     public void setChannelToken(int channelToken) {
-        this.channelToken = (byte) channelToken;
+        this.channelToken = channelToken;
     }
 
     @Override
